Actualizado logo

+++ PushUniversalWindows/Model/Smartphone.cs
@@ -5,6 +5,6 @@
         public string Id { get; set; }
         public string Modelo { get; set; }
         public string Fabricante { get; set; }
-        public decimal Precio { get; set; }
+        public double Precio { get; set; }
     }
 }
